--- a/Hazelcast.Benchmarks/Program.cs
+++ b/Hazelcast.Benchmarks/Program.cs
@@ -1,0 +1,12 @@
+﻿using BenchmarkDotNet.Running;
+
+namespace Hazelcast.Benchmarks
+{
+    class Program
+    {
+        public static void Main(string[] args)
+        {
+            var summary = BenchmarkRunner.Run(typeof(Program).Assembly);
+        }
+    }
+}