--- a/Hazelcast.Net.sln
+++ b/Hazelcast.Net.sln
@@ -14,9 +14,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		build-docs.bat = build-docs.bat
 		build-nuget.bat = build-nuget.bat
 		build.bat = build.bat
-		shared.project.props = shared.project.props
 		README.md = README.md
+		shared.project.props = shared.project.props
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hazelcast.Benchmarks", "Hazelcast.Benchmarks\Hazelcast.Benchmarks.csproj", "{C3004295-B687-402E-AB90-6B47B46E3153}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -36,6 +38,10 @@ Global
 		{0A545288-EF38-4594-9E53-65163B0C12CD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0A545288-EF38-4594-9E53-65163B0C12CD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0A545288-EF38-4594-9E53-65163B0C12CD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C3004295-B687-402E-AB90-6B47B46E3153}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C3004295-B687-402E-AB90-6B47B46E3153}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C3004295-B687-402E-AB90-6B47B46E3153}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C3004295-B687-402E-AB90-6B47B46E3153}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE