This PR changes the interface of `IFuture` and provides two different implementations for `synchronous` and `asynchronous` paths of execution.

Should be rebased/cherry picked after #256 is reviewed and merged.

### Benchmarks 
A synthetic benchmark written in a way to represent the flow behind the usage of a future. For the version before the changes, `SettableFuture<T>` is used for both synchronous and asynchronous scenarios. For the scenario after, `SyncFuture<T>` or `AsyncFuture<T>` are used according to the scenario.

#### Before:

|      Method |  Job | Runtime |     Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------ |----- |-------- |---------:|---------:|---------:|-------:|----------:|
|  SyncFuture |  Clr |     Clr | 685.1 ns | 13.66 ns | 37.84 ns | 0.0505 |     160 B |
| AsyncFuture |  Clr |     Clr | 873.1 ns | 16.79 ns | 20.62 ns | 0.0753 |     240 B |
|  SyncFuture | Core |    Core | 596.5 ns | 11.93 ns | 18.92 ns | 0.0477 |     152 B |
| AsyncFuture | Core |    Core | 574.0 ns | 11.39 ns | 24.27 ns | 0.0706 |     224 B |

#### After:

|      Method |  Job | Runtime |     Mean |     Error |   StdDev |  Gen 0 | Allocated |
|------------ |----- |-------- |---------:|----------:|---------:|-------:|----------:|
|  SyncFuture |  Clr |     Clr | 374.7 ns |  7.733 ns | 22.31 ns | 0.0072 |      24 B |
| AsyncFuture |  Clr |     Clr | 214.7 ns |  8.148 ns | 23.77 ns | 0.0584 |     184 B |
|  SyncFuture | Core |    Core | 576.3 ns | 11.318 ns | 22.60 ns | 0.0067 |      24 B |
| AsyncFuture | Core |    Core | 119.8 ns |  4.152 ns | 12.05 ns | 0.0532 |     168 B |