@devOpsHazelcast 2019-08-26
Linux Net Core Tests FAILed.