@devOpsHazelcast 2019-08-26
Windows Net Core Tests FAILed.