@devOpsHazelcast 2019-08-26
Windows .Net Framework Tests  PASSed.