@devOpsHazelcast 2019-08-27
Linux Net Core Tests PASSed.