@devOpsHazelcast 2019-08-27
Windows .Net Framework Tests  PASSed.