@devOpsHazelcast 2019-08-27
Windows Net Core Tests FAILed.