@zpqrtbnk 2020-12-08
We are about to release version 4.0 of the Hazelcast.Net client, and the code has been vastly refactored, rendering this PR... unfortunately irrelevant.

FYI, we're entirely moving towards pure C# async/await code... hopefully that will prove simpler to maintain, and fast.